+++ packages/aws-amplify-angular/src/components/authenticator/sign-in-component/sign-in.component.core.ts
@@ -51,7 +51,7 @@ const template = `
           required
           placeholder="{{ this.amplifyService.i18n().get('Enter your password') }}"
         />
-        <span class="amplify-form-action">{{ this.amplifyService.i18n().get('Forgot Password?') }}
+        <span class="amplify-form-action" *ngIf="!shouldHide('ForgotPassword')">{{ this.amplifyService.i18n().get('Forgot Password?') }}
         <a class="amplify-form-link"
             (click)="onForgotPassword()"
           >{{ this.amplifyService.i18n().get('Reset your password') }}</a></span>
@@ -62,7 +62,7 @@ const template = `
             (click)="onSignIn()"
           >{{ this.amplifyService.i18n().get('Sign In') }}</button>
         </div>
-        <div class="amplify-form-cell-left">
+        <div class="amplify-form-cell-left" *ngIf="!shouldHide('SignUp')">
           <div class="amplify-form-signup">
             {{ this.amplifyService.i18n().get('No account?') }}
             <a class="amplify-form-link" (click)="onSignUp()">
@@ -100,6 +100,13 @@ export class SignInComponentCore implements OnInit {
   }
 
   @Input()
+  set data(data: any) {
+    this.hide = data.hide ? data.hide : this.hide;
+  }
+
+  @Input() hide: string[] = [];
+
+  @Input()
   set authState(authState: AuthState) {
     this._authState = authState;
     this._show = includes(['signIn', 'signedOut', 'signedUp'], authState.state);
@@ -110,6 +117,11 @@ export class SignInComponentCore implements OnInit {
     if (!this.amplifyService.auth()){
       throw new Error('Auth module not registered on AmplifyService provider');
     }
+  }
+
+  shouldHide(comp) {
+    return this.hide.filter(item => item === comp)
+      .length > 0;
   }
 
   setUsername(username: string) {